backend/services/ticket_service/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
bcrypt
pydantic
email-validator
python-jose[cryptography]
